week4-api-microservices/python/requirements.txt
Django==4.0
django-cassandra-engine==1.6.3
python-decouple==3.5
djangorestframework==3.13.1 
